Bug fix
When saving a user, the '@JsonIgnore' annotation on line 22 was meant to hide the password field from the end-user. It worked but if you wanted to register a new user, Jackson would completely ignore that field. This was a big problem as the frontend endpoint that was calling this wouldn't persist a hashed (BCrypt) string into the MySQL Database. The new annotation '@JsonProperty' will only ignore the field if it was a 'GET' HTTP method because the 'access' parameter tells it to only allow writing for it to see it.